Generalized Postgres Import
