+++ src/lib/database/gamestat.ts
@@ -1,4 +1,4 @@
-import postgres from 'postgres'
+import sql from '$lib/database/postgres';
 // import { PGUSER, PGPASSWORD, PGHOST, PGPORT, PGDATABASE } from '$env/static/private'
 import type { GameStat } from '$lib/database/GameStats';
 
@@ -10,13 +10,6 @@ import type { GameStat } from '$lib/database/GameStats';
 //   database: PGDATABASE,
 // });
 
-const sql = postgres({
-  user: "postgres",
-  password: "test",
-  host: "localhost",
-  port: 5432,
-  database: "template1",
-});
 
 export async function getGameStats(): Promise<GameStat[]> {
   const players = await sql<GameStat[]>`
