+++ src/lib/database/postgres.ts
@@ -1,0 +1,11 @@
+import postgres from 'postgres'
+
+const sql = postgres({
+  user: "postgres",
+  password: "test",
+  host: "localhost",
+  port: 5432,
+  database: "template1",
+});
+
+export default sql
